requirements.txt
# alembic==1.14.0
# annotated-types==0.7.0
# anyio==4.6.2.post1
# async-timeout==5.0.0
# asyncclick==8.1.7.2
asyncpg==0.30.0
# bcrypt==4.0.1
# blinker==1.8.2
# certifi==2024.8.30
# charset-normalizer==3.4.1
# click==8.1.7
# colorama==0.4.6
# dnspython==2.7.0
# email_validator==2.2.0
# exceptiongroup==1.2.2
fastapi[standard]==0.115.6
# fastapi-cli==0.0.5
# fastapi-mail==1.4.1
# greenlet==3.1.1
# h11==0.14.0
# httpcore==0.16.3
# httptools==0.6.4
# httpx==0.23.3
# idna==3.10
# Jinja2==3.1.4
# Mako==1.3.6
# markdown-it-py==3.0.0
# MarkupSafe==3.0.2
# mdurl==0.1.2
# passlib==1.7.4
# pendulum==3.0.0
prometheus_client==0.21.1
# psycopg==3.2.3
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
# Pygments==2.18.0
# PyJWT==2.9.0
pynetgear==0.10.10
# python-dateutil==2.9.0.post0
# python-decouple==3.8
# python-dotenv==1.0.1
# python-multipart==0.0.17
# PyYAML==6.0.2
redis==5.2.1
# requests==2.32.3
# rfc3986==1.5.0
rich==13.9.4
RPi.GPIO==0.7.1
rtoml==0.11.0
# shellingham==1.5.4
# six==1.17.0
# sniffio==1.3.1
SQLAlchemy>=2.0.36
# starlette==0.41.2
# time-machine==2.16.0
# typer==0.12.5
# typing_extensions==4.12.2
# tzdata==2025.1
# urllib3==2.3.0
# uvicorn==0.32.0
# uvloop==0.21.0
# watchfiles==0.24.0
# websockets==13.1
